backend/requirements.txt
asgiref==3.8.1
Django==4.2.14
django-environ==0.11.2
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.01
django-cors-headers==4.4.0
requests==2.32.3
social-auth-app-django==5.4.2
django-allauth==65.0.2